requirements.txt
torch
ultralytics
opencv-python
torchvision
numpy
pillow
